feat: participant-specific config (only Electron)

+++ public/electron.js
@@ -261,6 +261,22 @@ ipc.on('save_video', (event, videoFileName, buffer) => {
   
 })
 
+ipc.on("save-config", (event, config, participantID, studyID) => {
+  if (savePath === "") {
+    savePath = getSavePath(participantID, studyID)
+  }
+
+  const configFileName = `pid_${participantID}_${today.getTime()}_config.json`
+  const saveConfigPath = getFullPath(configFileName)
+  if (!fs.existsSync(savePath)) {
+    fs.mkdir(savePath, { recursive: true }, (err) => {
+      log.error(err)
+      fs.writeFile(saveConfigPath, Buffer.from(JSON.stringify(config))).catch((error) => log.error(error))
+    })
+  } else {
+    fs.writeFile(saveConfigPath, Buffer.from(JSON.stringify(config))).catch((error) => log.error(error))
+  }
+})
 
 // EXPERIMENT END
 ipc.on('end', () => {
@@ -323,9 +339,13 @@ app.on('will-quit', () => {
     stream = false
 
     // copy file to config location
-    fs.mkdir(savePath, { recursive: true }, (err) => {
-      log.error(err)
+    if (!fs.existsSync(savePath)) {
+      fs.mkdir(savePath, { recursive: true }, (err) => {
+        log.error(err)
+        fs.copyFileSync(preSavePath, getFullPath(`pid_${participantID}_${today.getTime()}.json`))
+      })
+    } else {
       fs.copyFileSync(preSavePath, getFullPath(`pid_${participantID}_${today.getTime()}.json`))
-    })
+    }
   }
 })
